pooling to anydiff.Res
Pooling is kind of silly because it defeats the purpose of lazy sequences (since the pooled gradient has to be stored non-lazily). Still, it can be useful in cases where you feed a sequence to an RNN, then feed the output of that RNN to two or more different layers. This is the case in actor-critic models in RL.